+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Next
 
 ## New features
 
+- New option --print-srcaddr to show used IP source address in output (#461, thanks @gsnw-sebast)
 - New option --oiface for outgoing interface (#463, thanks @gsnw-sebast)
 
 ## Bugfixes and other changes